+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This is the changelog file of Elvvo.
 
 **Note:** All the dates below are in the IST time format.
 
+## <i>Version - 1.1.0 (7/12/2023)</i>
+
+1. There are two storage mechanisms for Elvvo: Web Storage and MySQL
+2. Added the following Python packages (dependencies): mysql-connector-python
+
 ## <i>Version - 1.0.0 (6/12/2023)</i>
 
 1. Uploaded "Elvvo" to GitHub for the first time.